Add ProductTab
weight: float -> double
complete .pro
add productMenu

+++ product.h
@@ -4,7 +4,7 @@
 struct Product{
 	int item;     // which kind of product,  not the id;
 	int category; // optional now;
-	float  weight;   // 
+	double  weight;   // 
 
 	int max_number[3]; // the maximum item number for small, 
 
@@ -12,12 +12,12 @@ public:
 	Product(){
 		item = -1;
 		category = -1;
-		float weight = 0.0;
+		double weight = 0;
 		max_number[0] = 0;
 		max_number[1] = 0;
 		max_number[2] = 0;
 	}
-	Product(int i, int c, float w, int number1, int number2, int number3){
+	Product(int i, int c, double w, int number1, int number2, int number3){
 		item = i;
 		category = c;
 		weight = w;
@@ -28,7 +28,7 @@ public:
 	
 	int getItem() {return item;}
 	int getCategory() {return category;}
-	float getWeight() {return weight;}
+	double getWeight() {return weight;}
 	int getNumber1() {return max_number[0];}
 	int getNumber2() {return max_number[1];}
 	int getNumber3() {return max_number[2];}
